code 09: Mood catalog documentation. - Part 05.
