+++ src/main/java/com/project/app_dog_back/application/service/IAnimoServiceImpl.java
@@ -1,6 +1,6 @@
 package com.project.app_dog_back.application.service;
 
-import com.project.app_dog_back.application.dto.AnimoDto;
+import com.project.app_dog_back.application.dto.catalog.AnimoDto;
 import com.project.app_dog_back.application.mapper.IAnimoMapper;
 import com.project.app_dog_back.domain.model.catalog.Animo;
 import com.project.app_dog_back.domain.repository.IAnimoRepository;
